+++ android/surge/src/main/java/co/instil/surge/player/SurgeRtspPlayerDelegate.java
@@ -11,15 +11,8 @@ package co.instil.surge.player;
  *
  */
 public interface SurgeRtspPlayerDelegate {
-
     /**
      * Called once a stream has stopped receiving frames from the RTSP stream.
      */
     void rtspPlayerDidTimeout();
-
-    /**
-     * Guaranteed to be called at most once per second with the current player frame rate.
-     * @param fps The Surge RTSP Player that called the delegate method.
-     */
-    void rtspPlayerDidUpdateFps(int fps);
 }